Ref. Integrations. Woocommerce gather to one class.
