+++ inc/cleantalk-public-integrations.php
@@ -271,297 +271,6 @@ function apbct_search_add_noindex()
 }
 
 /**
- * Test woocommerce checkout form for spam
- */
-function ct_woocommerce_checkout_check($_data, $errors)
-{
-    global $apbct, $cleantalk_executed;
-
-    if ( count($errors->errors) ) {
-        return;
-    }
-
-    if ( $apbct->settings['data__protect_logged_in'] == 0 && is_user_logged_in() ) {
-        do_action('apbct_skipped_request', __FILE__ . ' -> ' . __FUNCTION__ . '():' . __LINE__, $_POST);
-
-        return;
-    }
-
-    /**
-     * Filter for POST
-     */
-    $input_array = apply_filters('apbct__filter_post', $_POST);
-
-    //Getting request params
-    $ct_temp_msg_data = ct_get_fields_any($input_array);
-
-    $sender_email    = isset($ct_temp_msg_data['email']) ? $ct_temp_msg_data['email'] : '';
-    $sender_emails_array = isset($ct_temp_msg_data['emails_array']) ? $ct_temp_msg_data['emails_array'] : null;
-    $sender_nickname = isset($ct_temp_msg_data['nickname']) ? $ct_temp_msg_data['nickname'] : '';
-    $subject         = isset($ct_temp_msg_data['subject']) ? $ct_temp_msg_data['subject'] : '';
-    $message         = isset($ct_temp_msg_data['message']) ? $ct_temp_msg_data['message'] : array();
-
-    if ( $subject != '' ) {
-        $message = array_merge(array('subject' => $subject), $message);
-    }
-
-    $post_info = array();
-    $post_info['comment_type'] = 'order';
-    $post_info['post_url']     = Server::get('HTTP_REFERER');
-
-    $base_call_data = array(
-        'message'         => $message,
-        'sender_email'    => $sender_email,
-        'sender_nickname' => $sender_nickname,
-        'post_info'       => $post_info,
-        'sender_info'     => array('sender_url' => null, 'sender_emails_array' => $sender_emails_array)
-    );
-
-    //Making a call
-    $base_call_result = apbct_base_call($base_call_data);
-
-    if ( $apbct->settings['forms__wc_register_from_order'] ) {
-        $cleantalk_executed = false;
-    }
-
-    if ( isset($base_call_result['ct_result']) ) {
-        $ct_result = $base_call_result['ct_result'];
-
-        // Get request_id and save to static $hash
-        ct_hash($ct_result->id);
-
-        if ( $ct_result->allow == 0 ) {
-            if ( $apbct->settings['data__wc_store_blocked_orders'] ) {
-                apbct_woocommerce__store_blocked_order();
-            }
-            wp_send_json(array(
-                'result'   => 'failure',
-                'messages' => "<ul class=\"woocommerce-error\"><li>" . $ct_result->comment . "</li></ul>",
-                'refresh'  => 'false',
-                'reload'   => 'false'
-            ));
-        }
-    }
-}
-
-/**
- * @param \WC_Order $order
- * @return void
- * @throws \Automattic\WooCommerce\StoreApi\Exceptions\RouteException
- * @psalm-suppress UndefinedClass
- */
-function ct_woocommerce_checkout_check_from_rest($order)
-{
-    global $apbct, $cleantalk_executed;
-
-    if ( is_null($order) || ! ($order instanceof \WC_Order) ) {
-        return;
-    }
-
-    $sender_email    = $order->get_billing_email();
-    $sender_nickname = $order->get_billing_first_name() . ' ' . $order->get_billing_last_name();
-    $message         = $order->get_customer_note();
-
-    $post_info = array();
-    $post_info['comment_type'] = 'order';
-    $post_info['post_url']     = Server::get('HTTP_REFERER');
-
-    $base_call_data = array(
-        'message'         => $message,
-        'sender_email'    => $sender_email,
-        'sender_nickname' => $sender_nickname,
-        'post_info'       => $post_info,
-        'sender_info'     => array('sender_url' => null)
-    );
-
-    //Making a call
-    $base_call_result = apbct_base_call($base_call_data);
-
-    if ( $apbct->settings['forms__wc_register_from_order'] ) {
-        $cleantalk_executed = false;
-    }
-
-    if ( isset($base_call_result['ct_result']) ) {
-        $ct_result = $base_call_result['ct_result'];
-
-        // Get request_id and save to static $hash
-        ct_hash($ct_result->id);
-
-        if ( $ct_result->allow == 0 ) {
-            if ( $apbct->settings['data__wc_store_blocked_orders'] ) {
-                apbct_woocommerce__store_blocked_order();
-            }
-
-            if ( $order->get_status() === 'checkout-draft' ) {
-                try {
-                    $order->delete(true);
-                } catch (Exception $e) {
-                    error_log('Error deleting order: ' . $e->getMessage());
-                }
-            }
-
-            if ( class_exists('\Automattic\WooCommerce\StoreApi\Exceptions\RouteException') ) {
-                /** @psalm-suppress InvalidThrow */
-                throw new \Automattic\WooCommerce\StoreApi\Exceptions\RouteException(
-                    'woocommerce_store_api_checkout_order_processed',
-                    $ct_result->comment,
-                    403
-                );
-            }
-        }
-    }
-}
-
-/**
- * @return void
- *
- * @psalm-suppress UndefinedFunction
- */
-function apbct_woocommerce__store_blocked_order()
-{
-    global $wpdb;
-
-    $query = 'INSERT INTO ' . APBCT_TBL_WC_SPAM_ORDERS . ' (order_details, customer_details) 
-              VALUES (%s, %s) 
-              ON DUPLICATE KEY UPDATE order_details = %s, customer_details = %s';
-
-    $prepared_query = $wpdb->prepare($query, [
-        json_encode(wc()->session->cart),
-        json_encode($_POST),
-        json_encode(wc()->session->cart),
-        json_encode($_POST),
-    ]);
-
-    $wpdb->query($prepared_query);
-}
-
-/**
- * Save request_id for WC order
- * @param $order_id
- */
-function apbct_woocommerce__add_request_id_to_order_meta($order_id)
-{
-    $request_id = ct_hash();
-
-    if (!empty($request_id)) {
-        update_post_meta($order_id, 'cleantalk_order_request_id', sanitize_key($request_id));
-    }
-}
-
-/**
- * Triggered when adding an item to the shopping cart
- * for un-logged users
- *
- * @return bool
- * @psalm-suppress UnusedParam
- * @psalm-suppress UndefinedFunction
- */
-
-function apbct_wc__add_to_cart_unlogged_user()
-{
-    global $apbct;
-
-    $data = Post::get('data');
-    if (is_array($data) && isset($data['ct_bot_detector_event_token'])) {
-        $event_token = $data['ct_bot_detector_event_token'];
-    } elseif ( Get::get('ct_bot_detector_event_token') ) {
-        $event_token = Get::get('ct_bot_detector_event_token');
-    } else {
-        $event_token = null;
-    }
-
-    $message = apply_filters('apbct__filter_post', $_POST);
-
-    $post_info = array();
-    $post_info['comment_type'] = 'order__add_to_cart';
-    $post_info['post_url']     = Sanitize::cleanUrl(Server::get('HTTP_REFERER'));
-
-    if ( ! $apbct->stats['no_cookie_data_taken'] ) {
-        apbct_form__get_no_cookie_data();
-    }
-
-    //Making a call
-    $base_call_result = apbct_base_call(
-        array(
-            'message'     => $message,
-            'post_info'   => $post_info,
-            'js_on'       => apbct_js_test(Sanitize::cleanTextField(Cookie::get('ct_checkjs')), true),
-            'sender_info' => array('sender_url' => null),
-            'exception_action' => false,
-            'event_token' => $event_token,
-        )
-    );
-
-    if ( isset($base_call_result['ct_result']) ) {
-        $ct_result = $base_call_result['ct_result'];
-
-        if ( $ct_result->allow == 0 ) {
-            wc_add_notice($ct_result->comment, 'error');
-            return false;
-        }
-    }
-
-    return true;
-}
-
-/**
- * @param $add_to_cart_data
- * @param $request
- * @return mixed
- * @psalm-suppress UndefinedFunction
- */
-function apbct_wc_store_api_add_to_cart_data($add_to_cart_data, $request)
-{
-    global $apbct;
-
-    if ( ! $apbct->stats['no_cookie_data_taken'] && $request->get_param('ct_no_cookie_hidden_field') ) {
-        apbct_form__get_no_cookie_data(
-            ['ct_no_cookie_hidden_field' => $request->get_param('ct_no_cookie_hidden_field')],
-            false
-        );
-    }
-
-    return $add_to_cart_data;
-}
-
-/**
- * @param $customer
- * @param $request
- * @return void
- * @psalm-suppress UnusedParam
- */
-function apbct_wc_store_api_checkout_update_customer_from_request($customer, $request)
-{
-    global $apbct;
-
-    if ( ! $apbct->stats['no_cookie_data_taken'] ) {
-        apbct_form__get_no_cookie_data(
-            ['ct_no_cookie_hidden_field' => $request->get_param('ct_no_cookie_hidden_field')],
-            false
-        );
-    }
-}
-
-/**
- * @param $order
- * @return void
- * @throws \Automattic\WooCommerce\StoreApi\Exceptions\RouteException
- * @psalm-suppress UndefinedClass, UnusedParam, InvalidThrow
- */
-function apbct_wc_store_api_checkout_order_processed($order)
-{
-    global $ct_registration_error_comment;
-
-    if (class_exists('\Automattic\WooCommerce\StoreApi\Exceptions\RouteException')) {
-        throw new \Automattic\WooCommerce\StoreApi\Exceptions\RouteException(
-            'woocommerce_store_api_checkout_order_processed',
-            $ct_registration_error_comment,
-            403
-        );
-    }
-}
-
-/**
  * Public function - Tests for Pirate contact forms
  * return NULL
  */
@@ -1296,7 +1005,7 @@ function ct_registration_errors($errors, $sanitized_user_login = null, $user_ema
         $ct_registration_error_comment = $ct_result->comment;
 
         if (current_filter() === 'woocommerce_registration_errors') {
-            add_action('woocommerce_store_api_checkout_order_processed', 'apbct_wc_store_api_checkout_order_processed', 10, 2);
+            add_action('woocommerce_store_api_checkout_order_processed', ['Cleantalk\Antispam\IntegrationsByClass\Woocommerce', 'storeApiCheckoutOrderProcessed'], 10, 2);
         }
 
         if ( $buddypress === true ) {
@@ -3712,119 +3421,6 @@ function apbct_advanced_classifieds_directory_pro__check_register($response, $_f
     }
 
     return $response;
-}
-
-/*********** Woocommerce able add orders to spam ***********/
-
-/**
- * Register the new 'spam' status
- */
-add_filter('woocommerce_register_shop_order_post_statuses', 'apbct__wc_add_orders_spam_status');
-function apbct__wc_add_orders_spam_status($order_statuses)
-{
-    $order_statuses['wc-spamorder'] = array(
-        'label' => 'Spam',
-        'public' => false,
-        'exclude_from_search' => false,
-        'show_in_admin_all_list' => true,
-        'show_in_admin_status_list' => true,
-        'label_count' => _n_noop(
-            'Spam <span class="count">(%s)</span>',
-            'Spam <span class="count">(%s)</span>',
-            'cleantalk-spam-protect'
-        ),
-    );
-    return $order_statuses;
-}
-
-/**
- * Enable orders filtering by 'spam' status
- */
-add_filter('wc_order_statuses', 'apbct__wc_add_orders_spam_status_select'); // make spam selectable
-function apbct__wc_add_orders_spam_status_select($order_statuses)
-{
-    $order_statuses['wc-spamorder'] = 'Spam';
-    return $order_statuses;
-}
-
-/**
- * Don't show orders marked 'spam' in the common order list
- */
-add_action('parse_query', 'apbct__wc_add_orders_spam_status_hide_from_list'); // hide spam orders from total list
-function apbct__wc_add_orders_spam_status_hide_from_list($query)
-{
-    global $pagenow;
-
-    $query_vars = &$query->query_vars;
-
-    if ( $pagenow == 'edit.php'
-        && isset($query_vars['post_type'])
-        && $query_vars['post_type'] == 'shop_order'
-        && isset($query_vars['post_status'])
-        && is_array($query_vars['post_status'])
-        && ( $key = array_search('wc-spamorder', $query_vars['post_status']) ) !== false
-    ) {
-        unset($query_vars['post_status'][$key]);
-    }
-}
-
-/**
- * Add bulk actions: 'Mark as spam' and 'Unmark as spam'
- */
-add_filter('bulk_actions-edit-shop_order', 'apbct__wc_add_spam_action_to_bulk'); // add spam action to bulk
-function apbct__wc_add_spam_action_to_bulk($actions)
-{
-    if ( get_query_var('post_status') === 'wc-spamorder' ) {
-        $actions['unspamorder'] = __('Unmark as spam', 'cleantalk-spam-protect');
-    } else {
-        $actions['spamorder'] = __('Mark as spam', 'cleantalk-spam-protect');
-    }
-    return $actions;
-}
-
-/**
- * The bulk actions 'Mark as spam' and 'Unmark as spam' handler
- */
-add_filter('handle_bulk_actions-edit-shop_order', 'apbct__wc_add_spam_action_to_bulk_handle', 10, 3); // handle bulk action
-/**
- * @param $redirect
- * @param $action
- * @param $ids
- *
- * @return mixed|string
- * @psalm-suppress UndefinedClass
- */
-function apbct__wc_add_spam_action_to_bulk_handle($redirect, $action, $ids)
-{
-    if ( $action !== 'spamorder' &&  $action !== 'unspamorder' ) {
-        return $redirect;
-    }
-
-    // spam orders
-    $spam_ids = array();
-
-    foreach ($ids as $order_id) {
-        $order = new WC_Order((int)$order_id);
-        if ( $action === 'unspamorder' ) {
-            $order->update_status('wc-on-hold');
-        } else {
-            $spam_ids[] = $order_id;
-            $order->update_status('wc-spamorder');
-        }
-    }
-
-    // Send feedback to API
-    if (!empty($spam_ids)) {
-        apbct_woocommerce__orders_send_feedback($spam_ids);
-    }
-
-    return add_query_arg(
-        array(
-            'bulk_action' => 'marked_' . $action,
-            'changed' => count($ids),
-        ),
-        $redirect
-    );
 }
 
 function ct_mc4wp_hook($errors)
